마커 토클 이미지 변경, 코드 리팩토링
