+++ frontEnd/public/js/mapView/helperFunction.js
@@ -1,20 +1,17 @@
 import { isReleaseMode } from "../common/utils";
 import { distanceOption } from "./event";
+import { brandLogoMapper, ResetOptionButton, transStId, SelectStIdLogo, highhHsaleInfo, nohHsaleInfo, FillSTDetail, onlyhPrice, hnHhPrice} from "./ExtraFunctions";
+import { AdjustLocationToCurLocation } from "./SearchFunction";
+import { SetLPGmarker, SetMarkerofMapCenter, SetMarkerofResult, SetHGmarker, 
+    SetHmarker, SetGmarker, ResetMarkerArray, marker, markerArr, bringMarkerToFront} from "./marker";
+import { ShowDetailTab, appendNewChart, closeDetailTab, DetailTabDisplay, MoveSideBar, SideBarDisplay, ShowSideBar, HideSideBar} from "./DetailTab";
+import { getChartData, ShowChart} from "./chart";
 
 let map = "";
-let DetailTabDisplay = false;
 let searchOption = 0;
 let resultpoisData=[];
 let positionBounds;
-let marker, LPGmarker, HGmarker;
-let LPGmarkerArr = [];
-let HGmarkerArr = [];
-let markerArr = [];
 let ResultHtml = "";
-let SideBarDisplay = true;
-// let isOption1 = true;
-// let isOption2 = false; 
-let curzIndex = 101;
 
 function initTmap() {
 	map = new Tmapv3.Map("map_div", {
@@ -28,43 +25,23 @@ function initTmap() {
 	});
 
     map.on("ConfigLoad", function() {
-        let longitude = localStorage.getItem("longitude");
-        let latitude = localStorage.getItem("latitude");
-
-        if(longitude && latitude) {
-            FindAddressofSearchCoords(longitude, latitude);
-        }
-        else {
-            navigator.geolocation.getCurrentPosition((position) => {
-                FindAddressofSearchCoords(position.coords.longitude, position.coords.latitude);
-            });
-        }
-
-        const SortByHH = document.getElementById("sort_HH");
-        const SortByLL = document.getElementById("sort_LL");
-        const SortByGG = document.getElementById("sort_GG");
-
-        SortByHH.addEventListener('click', sortHhPrice);
-        SortByGG.addEventListener('click', sortGgPrice);
-        SortByLL.addEventListener('click', sortLlPrice);
+        MoveToUserCurLocation();
         
-        const MainFrame = document.querySelector(".main__ForMap");
-        const SideBarButton = MainFrame.querySelector(".main__SearchBarHideButton");
-        SideBarButton.addEventListener('click', (e) => MoveSideBar(e));
+        AddEventToOptionButton();
+        
+        AddEventToSideBarButton();
 
-        document.getElementsByClassName("main__GSTdetailCloseButton")[0].addEventListener('click', closeDetailTab);
+        AddEventToDetailTabButton();
+        
+        AddEventToResetButton();
 
-        // const SearchFromCurrentLocationButton = document.querySelector(".main__SearchFromCurrentLocation");
-        // SearchFromCurrentLocationButton.addEventListener('click', (e) => SelectSearchOption1(e));
-
-        const SearchFromRouteButton = document.querySelector(".main__SearchFromRoute");
-        SearchFromRouteButton.addEventListener('click', (e) => MoveToinitialPoint(e));
-
-        const AddOilingButton = document.getElementsByClassName("main__addOilingInfo")[0];
-        AddOilingButton.addEventListener('click', MakeOilingInformation);
+        AddEventToSearchButton();
     });
+}
 
+function AddEventToSearchButton() {
     const SearchButton = document.getElementById("main__btn_select");
+    SearchButton.addEventListener('click', SearchNearGasStation);
 
     map.on("DragStart", function() {
         SearchButton.style.opacity = "0";
@@ -72,49 +49,17 @@ function initTmap() {
     
     map.on("DragEnd", function() {
         SearchButton.style.opacity = "1";
-    });
+    });  
+}
 
-    SearchButton.addEventListener('click', SearchNearGasStation);
+function MoveMaptoSearchcoords(lon, lat){
+    var mapCenterPosition = new Tmapv3.LatLng(lat, lon);
+    map.setCenter(mapCenterPosition);
 }
 
 function FindAddressofSearchCoords(lon, lat) {
-    var mapCenterPosition = new Tmapv3.LatLng(lat, lon);
-    map.setCenter(mapCenterPosition);
-    $.ajax({
-        method : "GET",
-        url : "https://apis.openapi.sk.com/tmap/geo/reversegeocoding?version=1&format=json&callback=result&appKey=l7xx7250af6176574c63a12302edf09d020c",
-        async : false,
-        data : {
-            "coordType" : "WGS84GEO",
-            "addressType" : "A10",
-            "lon" : lon,
-            "lat" : lat
-        },
-        success : function(response) {
-            var arrResult = response.addressInfo;
-
-            var lastLegal = arrResult.legalDong.charAt(arrResult.legalDong.length - 1);
-
-            var newRoadAddr = arrResult.city_do + ' ' + arrResult.gu_gun + ' ';
-
-            if (arrResult.eup_myun == '' && (lastLegal == "읍" || lastLegal == "면")) {
-                newRoadAddr += arrResult.legalDong;
-            } 
-            else {
-                newRoadAddr += arrResult.eup_myun;
-            }
-            newRoadAddr += ' ' + arrResult.roadName + ' ' + arrResult.buildingIndex;
-
-            var result = "<div>" + newRoadAddr + "</div>";
-            const curAddr = document.getElementsByClassName("main__CurrentLocationAddress");
-            curAddr[0].innerHTML = result;
-        },
-        error : function(request, error) {
-            console.log("code:" + request.status + "\n"
-                    + "message:" + request.responseText + "\n"
-                    + "error:" + error);
-        }
-    });
+    MoveMaptoSearchcoords(lon, lat);
+    AdjustLocationToCurLocation(lon, lat);
 }
 
 function SearchNearGasStation(){
@@ -122,9 +67,7 @@ function SearchNearGasStation(){
         closeDetailTab();
     }
     searchOption=0;
-    $("#sort_GG").css("opacity", "0.5");
-    $("#sort_HH").css("opacity", "0.5");
-    $("#sort_LL").css("opacity", "0.5");
+    ResetOptionButton();
     $.ajax({
         method:"GET", // 요청 방식
         url:"https://apis.openapi.sk.com/tmap/pois/search/around?version=1&format=json&callback=result", 
@@ -142,7 +85,6 @@ function SearchNearGasStation(){
         },
         success:function(response) {
             FindAddressofSearchCoords(map.getCenter()._lng,map.getCenter()._lat);
-            console.log(response);
             if(response==null){
                 const NoResult = "검색 결과가 존재하지 않습니다.";
                 document.getElementById("searchResult").innerHTML = NoResult;
@@ -150,29 +92,18 @@ function SearchNearGasStation(){
             else {
                 resultpoisData = response.searchPoiInfo.pois.poi;
                 positionBounds = new Tmapv3.LatLngBounds();
-                ResetMarkerArray(LPGmarkerArr);
-                ResetMarkerArray(HGmarkerArr);
                 ShowResult(resultpoisData, positionBounds); // 거리순 결과 데이터출력
             }
         },
         error:function(request, error){
-            console.log("code:"+request.status+"\n"+"message:"+request.responseText+"\n"+"error:"+error);
+            console.log("code:"+request.status + ", message:" + request.responseText + ", error:"+error);
         }
     });
 }
 
-function ResetMarkerArray(markerArr) {
-    if(markerArr.length > 0){
-        for(var i in markerArr){
-            markerArr[i].setMap(null);
-        }
-        markerArr = [];
-    }
-}
+
 
 function ShowResult(ResultArray, positionBounds){
-    ResetMarkerArray(HGmarkerArr);
-    ResetMarkerArray(LPGmarkerArr);
     ResetMarkerArray(markerArr);
     ResultHtml = "";
     const resultArea = document.querySelector(".main__SearchResult");
@@ -181,61 +112,29 @@ function ShowResult(ResultArray, positionBounds){
         if(ResultArray[k].hhPrice==0 && ResultArray[k].ggPrice==0 && ResultArray[k].llPrice==0){
             continue;
         }
-        var noorLat = Number(ResultArray[k].noorLat);
-        var noorLon = Number(ResultArray[k].noorLon);
 
         ResultHtml += SelectStIdLogo(ResultArray[k].stId, ResultArray[k].name, ResultArray[k].radius);
-        var pointCng = new Tmapv3.Point(noorLon, noorLat);
-        var projectionCng = new Tmapv3.Projection.convertEPSG3857ToWGS84GEO(pointCng);
-        
-        var lat = projectionCng._lat;
-        var lon = projectionCng._lng;
 
-        var markerPosition = new Tmapv3.LatLng(lat, lon);
-
-        marker = new Tmapv3.Marker({
-            position : map.getCenter(),
-            icon : "http://tmapapi.sktelecom.com/upload/tmap/marker/pin_b_m_m.png",
-            map:map
-            });
-        markerArr.push(marker);
+        var markerPosition = SetMarkerofResult(ResultArray[k]);
+        SetMarkerofMapCenter();
 
         if(ResultArray[k].llPrice != 0){
-            LPGmarker = new Tmapv3.Marker({
-                position : markerPosition,
-                iconHTML: `<div class='Map_Marker_LPG' id='${ResultArray[k].name}'><img class='img_LPG'><span id="LL">${ResultArray[k].llPrice}</span></div>`,
-                iconSize : Tmapv3.Size(10, 20),
-                map:map
-            });
-            markerArr.push(LPGmarker);
+            SetLPGmarker(markerPosition, ResultArray[k]);
         }
         else {
-            HGmarker = new Tmapv3.Marker({
-                position : markerPosition,
-                iconHTML: `<div class='Map_Marker_HG' id='${ResultArray[k].name}'><img class='img_HG'><span id="H">${ResultArray[k].hhPrice}</span>
-                            <span id='G'>${ResultArray[k].ggPrice}</span></div>`,
-                iconSize : Tmapv3.Size(10, 20),
-                map:map
-            });
-            markerArr.push(HGmarker);
+            SetHGmarker(markerPosition, ResultArray[k]);
         }
+
         if(ResultArray[k].highHhSale != 0){
-            ResultHtml += `<div class='main__ResultList__Contents'>
-                        휘발유: <span> ${ResultArray[k].hhPrice == 0 ? '-' : ResultArray[k].hhPrice} </span>
-                        <i class="fa-solid fa-circle-check"></i>&nbsp;
-                        경유: <span> ${ResultArray[k].ggPrice == 0 ? '-' : ResultArray[k].ggPrice} </span>&nbsp;&nbsp;
-                        LPG: <span> ${ResultArray[k].llPrice == 0 ? '-' : ResultArray[k].llPrice} </span></div></div></div>`;
+            ResultHtml += highhHsaleInfo(ResultArray[k]);
 
         }
-        else {
-            ResultHtml += `<div class='main__ResultList__Contents'>
-                        휘발유: <span> ${ResultArray[k].hhPrice == 0 ? '-' : ResultArray[k].hhPrice} </span>&nbsp;&nbsp;
-                        경유: <span> ${ResultArray[k].ggPrice == 0 ? '-' : ResultArray[k].ggPrice} </span>&nbsp;&nbsp;
-                        LPG: <span> ${ResultArray[k].llPrice == 0 ? '-' : ResultArray[k].llPrice} </span></div></div></div>`;
-
+        else { 
+            ResultHtml += nohHsaleInfo(ResultArray[k]);
         }
         positionBounds.extend(markerPosition);
     }
+
     addmouseOverEventtoMarker();
     addClickEventToMarker(ResultArray);
     resultArea.innerHTML = ResultHtml;
@@ -243,43 +142,6 @@ function ShowResult(ResultArray, positionBounds){
     map.fitBounds(positionBounds);	
 }
 
-const brandLogoMapper = (brandName) => {
-    switch(brandName) {
-        case "알뜰":
-            return "https://team6-public-image.s3.ap-northeast-2.amazonaws.com/주유소+로고_0213/알뜰.png";
-        case "오일뱅크":
-            return "https://team6-public-image.s3.ap-northeast-2.amazonaws.com/주유소+로고_0213/오일뱅크.png";
-        case "E1":
-            return "https://team6-public-image.s3.ap-northeast-2.amazonaws.com/주유소+로고_0213/E1.png";
-        case "ex-OIL":
-            return "https://team6-public-image.s3.ap-northeast-2.amazonaws.com/주유소+로고_0213/ex.png";
-        case "ex":
-            return "https://team6-public-image.s3.ap-northeast-2.amazonaws.com/주유소+로고_0213/ex.png";
-        case "GS":
-            return "https://team6-public-image.s3.ap-northeast-2.amazonaws.com/주유소+로고_0213/GS.png";
-        case "S-Oil":
-            return "https://team6-public-image.s3.ap-northeast-2.amazonaws.com/주유소+로고_0213/S-Oil.png";
-        case "SK":
-            return "https://team6-public-image.s3.ap-northeast-2.amazonaws.com/주유소+로고_0213/SK_Gas.png";
-        case "NH-OIL":
-            return "https://team6-public-image.s3.ap-northeast-2.amazonaws.com/주유소+로고_0213/NH-OIL.png";
-    }
-
-    return "./img/GasStation_Image/custom.png";
-}
-
-function SelectStIdLogo(stId, name, radius){
-    return `
-        <div class='main__ResultList'> 
-            <img class='main__ResultList__Title_Logo' src = ${brandLogoMapper(stId)}>
-            <div class='main__ResultList_TitlenContents'>
-                <div class='main__ResultList__Title'>  
-                    <div id="stName">${name}</div>
-                    <div id="stRadius">${Math.round(radius*10)/10}km</div>
-                </div>
-                
-    `;
-}
 
 function addEventToResult(ResultArray){
     const GSTArray = document.getElementsByClassName("main__ResultList");
@@ -292,19 +154,10 @@ function addEventToResult(ResultArray){
 
 function ShowGSTDetail(event, ResultArray){
     map.setZoom(17);
-    if(DetailTabDisplay==false){
-        const GSTDetailTab = document.getElementsByClassName("main__GSTDetailTab");
-        GSTDetailTab[0].style.marginLeft = "0vw";
-        const SideSearchBar = document.getElementsByClassName("main__SideSearchBar");
-        SideSearchBar[0].style.borderRadius = "0";
-        DetailTabDisplay = true;
-    }
 
-    document.getElementById('myChart').remove();
-    const newChart = document.createElement('canvas');
-    newChart.id= 'myChart'; 
+    if(DetailTabDisplay==false) ShowDetailTab();
+    appendNewChart();
 
-    document.getElementsByClassName('main__GSTdetail__Contents__Chart')[0].append(newChart);
     const ResultTitle = event.target.closest(".main__ResultList");
     const SelectedTitle = ResultTitle.querySelector("#stName");
 
@@ -313,16 +166,7 @@ function ShowGSTDetail(event, ResultArray){
             break;
         }
     }
-    var noorLat = Number(ResultArray[k].noorLat);
-    var noorLon = Number(ResultArray[k].noorLon);
-
-    var pointCng = new Tmapv3.Point(noorLon, noorLat);
-    var projectionCng = new Tmapv3.Projection.convertEPSG3857ToWGS84GEO(pointCng);
-    
-    var lat = projectionCng._lat;
-    var lon = projectionCng._lng;
-
-    var markerPosition = new Tmapv3.LatLng(lat, lon);
+    var markerPosition = SetMarkerofResult(ResultArray[k]);
 
     if(searchOption==0){
         ShowOnlyCurMarker(SelectedTitle.innerHTML);
@@ -336,140 +180,8 @@ function ShowGSTDetail(event, ResultArray){
     map.zoomIn();
 
     FillSTDetail(ResultArray[k]);
-    const HOST_URL = isReleaseMode() ? 
-                        "https://www.oilpocket.kro.kr" :
-                        "http://localhost:8080" ;
-
-    $.ajax({
-        method:"GET",
-        url:`${HOST_URL}/api/v1/gas-station/` + ResultArray[k].name + "/" + ResultArray[k].roadName + "/" 
-        + ResultArray[k].buildingNo1 + "/" + transStId(ResultArray[k].stId) + "/month",
-        xhrFields: {
-            withCredentials: true
-        },
-        success:function(response) {
-            // console.log(transStId(ResultArray[k].stId));
-            // console.log(ResultArray[k].roadName);
-            // console.log(ResultArray[k].buildingNo1);
-            // console.log(response);
-            ShowChart(response);
-        },
-        error:function(request, error){
-            console.log("code:"+request.status+"\n"+"message:"+request.responseText+"\n"+"error:"+error);
-        }
-    });
-}
-
-
-function FillSTDetail(ResultArrayElem){
-
-    const AddOilingButton = document.getElementsByClassName("main__addOilingInfo")[0];
-    if(searchOption == 0){
-        AddOilingButton.style.display = 'none';
-    }
-    else {
-        AddOilingButton.style.display = 'block';
-    }
-    const ST_name = document.getElementById("GSTdetail__Name");
-    ST_name.innerHTML = ResultArrayElem.name;
-
-    const ST_Price = document.getElementsByClassName("main__GSTdetail__Contents__Oilprice");
-    var PriceinnerHtml = "";
-    PriceinnerHtml += "<i class='fa-solid fa-gas-pump'></i>";
-    if(ResultArrayElem.hhPrice != 0){
-        PriceinnerHtml += "휘발유 : " + ResultArrayElem.hhPrice + "&nbsp&nbsp";
-    }
-    if(ResultArrayElem.ggPrice != 0) {
-        PriceinnerHtml += "경유 : " + ResultArrayElem.ggPrice; + "&nbsp&nbsp";
-    }
-    if(ResultArrayElem.llPrice != 0) {
-        PriceinnerHtml += "LPG : " + ResultArrayElem.llPrice;
-    }
-    ST_Price[0].innerHTML = PriceinnerHtml;
-
-    const ST_addr = document.getElementById("GSTdetail_addr");
-    ST_addr.innerHTML = `${ResultArrayElem.upperAddrName} ${ResultArrayElem.middleAddrName} ${ResultArrayElem.lowerAddrName}
-                        ${ResultArrayElem.detailAddrName} ${ResultArrayElem.firstNo}
-                        ( ${ResultArrayElem.roadName} ${ResultArrayElem.buildingNo1} )`
-
-    const ST_phone = document.getElementById("GSTdetail_phone");
-    var telNolen = ResultArrayElem.telNo.length;
-    if(ResultArrayElem.telNo.includes('02')) {
-        var psTelNo = ResultArrayElem.telNo.slice(0, 2) + '-' + ResultArrayElem.telNo.slice(2, telNolen-4) + '-' + ResultArrayElem.telNo.slice(telNolen-4, telNolen);
-    }
-    else {
-        var psTelNo = ResultArrayElem.telNo.slice(0, 3) + '-' + ResultArrayElem.telNo.slice(3, telNolen-4) + '-' + ResultArrayElem.telNo.slice(telNolen-4, telNolen);
-    }
-
-    ST_phone.innerHTML = psTelNo;
-
-    const ST_url = document.getElementById("GSTdetail_url");
-    ST_url.innerHTML = `https://www.${ResultArrayElem.stId}.com`;
-
-    const ST_PriceTable = document.getElementsByClassName("main__GSTdetail__Contents__OilPriceTable__Contents");
-    ST_PriceTable[0].innerHTML = "";
-    if(ResultArrayElem.hhPrice != 0){
-        const ST_OilPrice = document.createElement("div");
-        ST_OilPrice.classList.add("main__GSTdetail__Contents__OilPriceTable__Contents__Row");
-        ST_OilPrice.innerHTML = `<div class="main__GSTdetail__Contents__OilPriceTable__Contents__Oil">
-                                    휘발유
-                                    </div>
-                                    <div class="main__GSTdetail__Contents__OilPriceTable__Contents__Price">
-                                        <span id="Contents_HH">${ResultArrayElem.hhPrice}</span>
-                                    </div>`
-        ST_PriceTable[0].append(ST_OilPrice);
-    }
     
-    if(ResultArrayElem.ggPrice != 0) {
-        const ST_OilPrice = document.createElement("div");
-        ST_OilPrice.classList.add("main__GSTdetail__Contents__OilPriceTable__Contents__Row");
-        ST_OilPrice.innerHTML = `<div class="main__GSTdetail__Contents__OilPriceTable__Contents__Oil">
-                                    경유
-                                    </div>
-                                    <div class="main__GSTdetail__Contents__OilPriceTable__Contents__Price">
-                                        <span id="Contents_HH">${ResultArrayElem.ggPrice}</span>
-                                    </div>`
-        ST_PriceTable[0].append(ST_OilPrice);
-    }
-
-    if(ResultArrayElem.highHhPrice != 0) {
-        const ST_OilPrice = document.createElement("div");
-        ST_OilPrice.classList.add("main__GSTdetail__Contents__OilPriceTable__Contents__Row");
-        ST_OilPrice.innerHTML = `<div class="main__GSTdetail__Contents__OilPriceTable__Contents__Oil">
-                                    고급 휘발유
-                                    </div>
-                                    <div class="main__GSTdetail__Contents__OilPriceTable__Contents__Price">
-                                        <span id="Contents_HH">${ResultArrayElem.highHhPrice}</span>
-                                    </div>`
-        ST_PriceTable[0].append(ST_OilPrice);
-    }
-
-    if(ResultArrayElem.llPrice != 0) {
-        const ST_OilPrice = document.createElement("div");
-        ST_OilPrice.classList.add("main__GSTdetail__Contents__OilPriceTable__Contents__Row");
-        ST_OilPrice.innerHTML = `<div class="main__GSTdetail__Contents__OilPriceTable__Contents__Oil">
-                                    LPG
-                                    </div>
-                                    <div class="main__GSTdetail__Contents__OilPriceTable__Contents__Price">
-                                        <span id="Contents_HH">${ResultArrayElem.llPrice}</span>
-                                    </div>`
-        ST_PriceTable[0].append(ST_OilPrice);
-    }
-}
-
-function transStId(stId){
-    switch(stId){
-        case 'S-Oil': return 'S-OIL';
-        case '오일뱅크' : return '현대오일뱅크';
-        case '알뜰' : return '알뜰주유소';
-        case 'ex-OIL' : return '알뜰(ex)';
-        case 'NH-OIL' : return 'NH-OIL';
-        case 'SK' : return 'SK'; 
-        case 'GS' : return 'GS칼텍스';
-        case '자가상표' : return '자가상표';
-        case 'E1' : return 'E1';
-    }
-    return '자가상표';
+    getChartData(ResultArray[k]);
 }
 
 function sortHhPrice() {
@@ -485,7 +197,6 @@ function sortHhPrice() {
         let HHsort = resultpoisData.sort(function(a, b) {
             return a.hhPrice - b.hhPrice;
         })
-        //const sortlen = HHsort.length;
         for(var k=0;k<HHsort.length;k++){
             if(HHsort[k].hhPrice != 0){
                 break;
@@ -532,7 +243,6 @@ function sortLlPrice() {
         let LLsort = resultpoisData.sort(function(a, b) {
             return a.llPrice - b.llPrice;
         })
-        //const sortlen = HHsort.length;
         for(var k=0;k<LLsort.length;k++){
             if(LLsort[k].llPrice != 0){
                 break;
@@ -546,70 +256,36 @@ function ShowResultByOption(ResultArray) {
     ResultHtml = "";
     const resultArea = document.querySelector(".main__SearchResult");
     resultArea.innerHTML = "";
-    ResetMarkerArray(HGmarkerArr);
-    ResetMarkerArray(LPGmarkerArr);
     ResetMarkerArray(markerArr);
     var positionBounds = new Tmapv3.LatLngBounds();
-    var i =0;
+    var loopCounter = 10 > ResultArray.length ? ResultArray.length : 10;
 
-    for(var k=0;k<ResultArray.length;k++){      
-        if(i > 9) break;
-        
-        var noorLat = Number(ResultArray[k].noorLat);
-        var noorLon = Number(ResultArray[k].noorLon);
-
+    for(var k=0;k<loopCounter;k++){      
         ResultHtml += SelectStIdLogo(ResultArray[k].stId, ResultArray[k].name, ResultArray[k].radius);
-        var pointCng = new Tmapv3.Point(noorLon, noorLat);
-        var projectionCng = new Tmapv3.Projection.convertEPSG3857ToWGS84GEO(pointCng);
-        
-        var lat = projectionCng._lat;
-        var lon = projectionCng._lng;
-
-        var markerPosition = new Tmapv3.LatLng(lat, lon);
+        var markerPosition = SetMarkerofResult(ResultArray[k]);
 
         if(searchOption==1){
             if(ResultArray[k].highHhSale != 0){
-                ResultHtml += `<div class='main__ResultList__Contents'>
-                        휘발유: <span>${ResultArray[k].hhPrice == 0 ? '-' : ResultArray[k].hhPrice}</span>
-                        <i class="fa-solid fa-circle-check"></i>
-                        </div></div></div>`;
+                ResultHtml += hnHhPrice(ResultArray[k]);
             }
             else {
-                ResultHtml += `<div class='main__ResultList__Contents'>
-                        휘발유: <span>${ResultArray[k].hhPrice == 0 ? '-' : ResultArray[k].hhPrice}</span>
-                        </div></div></div>`;
+                ResultHtml += onlyhPrice(ResultArray[k]);
             }
-            marker = new Tmapv3.Marker({
-                position : markerPosition,
-                iconHTML: `<div class='Map_Marker_hprice' id='${ResultArray[k].name}'><img class='img_HnG'><span id="Price">${ResultArray[k].hhPrice}</span></div>`,
-                iconSize : Tmapv3.Size(10, 20),
-                map:map
-            }); 
+            SetHmarker(markerPosition, ResultArray[k]);
         }
         else if(searchOption==2){
             ResultHtml += `<div class='main__ResultList__Contents'>
                         경유: <span> ${ResultArray[k].ggPrice == 0 ? '-' : ResultArray[k].ggPrice} </span>&nbsp;&nbsp;
                         </div></div></div>`;
-            marker = new Tmapv3.Marker({
-                position : markerPosition,
-                iconHTML: `<div class='Map_Marker_gprice' id='${ResultArray[k].name}'><img class='img_HnG'><span id="Price">${ResultArray[k].ggPrice}</span></div>`,
-                iconSize : Tmapv3.Size(10, 20),
-                map:map
-            });
+            SetGmarker(markerPosition, ResultArray[k]);
         }
         else if(searchOption==3){
             ResultHtml += `<div class='main__ResultList__Contents'>
                         LPG: <span> ${ResultArray[k].llPrice == 0 ? '-' : ResultArray[k].llPrice} </span>&nbsp;&nbsp;
                         </div></div></div>`;
-            marker = new Tmapv3.Marker({
-                position : markerPosition,
-                iconHTML: `<div class='Map_Marker_LPG' id='${ResultArray[k].name}'><img class='img_LPG'><span id="LL">${ResultArray[k].llPrice}</span></div>`,
-                iconSize : Tmapv3.Size(10, 20),
-                map:map
-            });
+            SetLPGmarker(markerPosition, ResultArray[k]);
         }
         markerArr.push(marker);
-        i++;
         positionBounds.extend(markerPosition);
     }
     addmouseOverEventtoMarker();
@@ -617,156 +293,6 @@ function ShowResultByOption(ResultArray) {
     resultArea.innerHTML = ResultHtml;
     addEventToResult(ResultArray);
     map.fitBounds(positionBounds);	
-}
-
-function MoveSideBar(e){
-    const HideSideBar = document.getElementsByClassName("main__SearchBarnDetailTab");
-    const HideSideBarButton = e.target.closest(".main__SearchBarHideButton");
-    if(SideBarDisplay){
-        if(DetailTabDisplay){
-            HideSideBar[0].style.marginLeft = "-44vw";
-        }
-        else {
-            HideSideBar[0].style.marginLeft = "-22vw";
-        }
-        SideBarDisplay = false;
-        HideSideBarButton.innerHTML = "<img src='img/🦆 icon _chevron right_.svg'>";
-        
-    }
-    else {
-        HideSideBar[0].style.marginLeft = "0vw";
-        SideBarDisplay = true;
-        HideSideBarButton.innerHTML = "<img src='img/🦆 icon _chevron left_.svg'>";
-    }
-}
-
-function closeDetailTab() {
-    const GSTDetailTab = document.getElementsByClassName("main__GSTDetailTab");
-    GSTDetailTab[0].style.marginLeft = "-22vw";
-    const SideSearchBar = document.getElementsByClassName("main__SideSearchBar");
-    SideSearchBar[0].style.borderRadius = "0 5px 5px 0";
-    DetailTabDisplay = false;
-}
-
-function ShowChart(response){
-    const dateLabels = [];
-    const PREGasolineData = [];
-    const GasolineData = [];
-    const DiselData = [];
-    const LPGData = [];
-    let data;
-    const PriceInformation = response.data.details;
-    for(var k in PriceInformation){
-        if(PriceInformation[k].gasType == 'PREMIUM_GASOLINE'){
-            PREGasolineData.push(PriceInformation[k].price);
-        }
-        else if(PriceInformation[k].gasType == 'GASOLINE'){
-            GasolineData.push(PriceInformation[k].price);
-        }
-        else if(PriceInformation[k].gasType == 'DIESEL'){
-            DiselData.push(PriceInformation[k].price);
-        }
-        else if(PriceInformation[k].gasType == 'LPG'){
-            LPGData.push(PriceInformation[k].price);
-        }
-    }
-    console.log("여기까지는 되냐?");
-    var Preflag=0;
-    for(var index=0; index<30;index++){
-        if(PREGasolineData[index]==0){
-            continue;
-        }
-        else Preflag=1;
-    }
-
-    if(GasolineData.length == 0){
-        for(var k =0;k<30;k++){
-            dateLabels.push(PriceInformation[k].date.slice(-2));
-        }
-        data = {
-            labels: dateLabels,
-            datasets: [
-              {
-                label: 'LPG',
-                backgroundColor: '#008000',
-                borderColor: '#008000',
-                data: LPGData,
-                fill:false,
-              }
-          ]
-        };
-    }
-    else {
-        if(Preflag ==0){
-            for(var k =0;k<90; k+=3){
-                dateLabels.push(PriceInformation[k].date.slice(-2));
-            }
-            data = {
-                labels: dateLabels,
-                datasets: [
-                  {
-                      label: '휘발유',
-                      backgroundColor: '#93bf85',
-                      borderColor: '#93bf85',
-                      data: GasolineData,
-                      fill:false,
-                  },
-                  {
-                      label: '경유',
-                      backgroundColor: '#008000',
-                      borderColor: '#008000',
-                      data: DiselData,
-                      fill:false,
-                  }
-              ]
-            };
-        }
-        else {
-            for(var k =0;k<90;k+=3){
-                dateLabels.push(PriceInformation[k].date.slice(-2));
-            }
-            data = {
-                labels: dateLabels,
-                datasets: [
-                  {
-                      label: '고급유',
-                      backgroundColor: '#dbead5',
-                      borderColor: '#dbead5',
-                      data: PREGasolineData,
-                      fill:false,
-                  },
-                  {
-                      label: '휘발유',
-                      backgroundColor: '#93bf85',
-                      borderColor: '#93bf85',
-                      data: GasolineData,
-                      fill:false,
-                  },
-                  {
-                      label: '경유',
-                      backgroundColor: '#008000',
-                      borderColor: '#008000',
-                      data: DiselData,
-                      fill:false,
-                  }
-              ]
-            };
-        }
-    }
-
-    const config = {
-      type: 'line',
-      data: data,
-      options: {
-        responsive: false,
-        pointStyle:false,
-        }
-    }
-
-    const myChart = new Chart(
-      document.getElementById('myChart'),
-      config
-    )
 }
 
 function ShowOnlyCurMarker(SelectedTitle){
@@ -786,7 +312,24 @@ function ShowOnlyCurMarker(SelectedTitle){
 function ToggleCurMarker(SelectedTitle){
     const selectedmarker = document.getElementById(SelectedTitle);
 
-    if(searchOption==1){
+    if(searchOption==0){
+        const HGmarkerARR = document.getElementsByClassName('Map_Marker_HG');
+        for(var k=0;k<HGmarkerARR.length;k++){
+            HGmarkerARR[k].classList.remove('selected__Map_Marker_HG');
+        }
+        const LmarkerARR = document.getElementsByClassName('Map_Marker_LPG');
+        for(var k=0;k<LmarkerARR.length;k++){
+            LmarkerARR[k].classList.remove('selected__Map_Marker_LPG');
+        }
+
+        if(SelectedTitle.includes('주유소')){
+            selectedmarker.classList.toggle('selected__Map_Marker_HG');
+        }
+        else if(SelectedTitle.includes('충전소')){
+            selectedmarker.classList.toggle('selected__Map_Marker_LPG');
+        }
+    }
+    else if(searchOption==1){
         const HmarkerARR = document.getElementsByClassName('Map_Marker_hprice');
         for(var k=0;k<HmarkerARR.length;k++){
             HmarkerARR[k].classList.remove('selected__Map_Marker_hprice');
@@ -821,13 +364,6 @@ function addmouseOverEventtoMarker() {
     for(var k=0;k<HnGmapmarkerarr.length;k++){
         HnGmapmarkerarr[k].addEventListener('mouseover', (e) => bringMarkerToFront(e));
     }
-}
-
-function bringMarkerToFront(e) {
-    const MouseonMarker = e.target.closest('.vsm-marker');
-    // console.log(MouseonMarker);
-    MouseonMarker.style.zIndex = `${curzIndex}`;
-    curzIndex++;
 }
 
 function addClickEventToMarker(ResultArray){
@@ -882,72 +418,26 @@ function ShowGSTDetailByMarker(e, ResultArray){
 
 
 function ShowDetail(ResultArray, markername){
-    //map.setZoom(17);
-
+    if(SideBarDisplay==false){
+        ShowSideBar();
+    }
     if(DetailTabDisplay==false){
-        const GSTDetailTab = document.getElementsByClassName("main__GSTDetailTab");
-        GSTDetailTab[0].style.marginLeft = "0vw";
-        const SideSearchBar = document.getElementsByClassName("main__SideSearchBar");
-        SideSearchBar[0].style.borderRadius = "0";
-        DetailTabDisplay = true;
+        ShowDetailTab();
     }
 
-    document.getElementById('myChart').remove();
-    const newChart = document.createElement('canvas');
-    newChart.id= 'myChart'; 
-    document.getElementsByClassName('main__GSTdetail__Contents__Chart')[0].append(newChart);
+    appendNewChart();
     
-    //console.log(HGmarkerGSTname.id);
     for(var k=0;k<ResultArray.length;k++){
         if(ResultArray[k].name == markername.id){
             break;
         }
     }
-    // var noorLat = Number(ResultArray[k].noorLat);
-    // var noorLon = Number(ResultArray[k].noorLon);
-
-    // var pointCng = new Tmapv3.Point(noorLon, noorLat);
-    // var projectionCng = new Tmapv3.Projection.convertEPSG3857ToWGS84GEO(pointCng);
-    
-    // var lat = projectionCng._lat;
-    // var lon = projectionCng._lng;
-
-    // var markerPosition = new Tmapv3.LatLng(lat, lon);
-
-    // if(searchOption==0){
-    //     ShowOnlyCurMarker(markername.id);
-    // }
-    // else {
-    //     ToggleCurMarker(markername.id);
-    // }
-
-    // map.setCenter(markerPosition);
-
-    // map.zoomIn();
-
-    // console.log(ResultArray[k]);
+    ToggleCurMarker(markername.id);
     FillSTDetail(ResultArray[k]);
-    const HOST_URL = isReleaseMode() ? 
-                        "https://www.oilpocket.kro.kr" :
-                        "http://localhost:8080" ;
-
-    $.ajax({
-        method:"GET",
-        url:`${HOST_URL}/api/v1/gas-station/` + ResultArray[k].name + "/" + ResultArray[k].roadName + "/" 
-        + ResultArray[k].buildingNo1 + "/" + transStId(ResultArray[k].stId) + "/month",
-        xhrFields: {
-            withCredentials: true
-        },
-        success:function(response) {
-            ShowChart(response);
-        },
-        error:function(request, error){
-            console.log("code:"+request.status+"\n"+"message:"+request.responseText+"\n"+"error:"+error);
-        }
-    });
+    getChartData(ResultArray[k]);
 }
 
-function MoveToinitialPoint(e){
+function MoveToinitialPoint(){
     ResetMarkerArray(markerArr);
     let longitude = localStorage.getItem("longitude");
     let latitude = localStorage.getItem("latitude");
@@ -963,13 +453,48 @@ function MoveToinitialPoint(e){
     const resultArea = document.querySelector(".main__SearchResult");
     resultArea.innerHTML = "";
 
-    marker = new Tmapv3.Marker({
-        position : map.getCenter(),
-        icon : "http://tmapapi.sktelecom.com/upload/tmap/marker/pin_b_m_m.png",
-        map:map
-        });
-    markerArr.push(marker);
+    SetMarkerofMapCenter();
 }
-export { initTmap }
+
+
+function MoveToUserCurLocation(){
+    let longitude = localStorage.getItem("longitude");
+    let latitude = localStorage.getItem("latitude");
+
+    if(longitude && latitude) {
+        FindAddressofSearchCoords(longitude, latitude);
+    }
+    else {
+        navigator.geolocation.getCurrentPosition((position) => {
+            FindAddressofSearchCoords(position.coords.longitude, position.coords.latitude);
+        });
+    }
+}
+
+function AddEventToOptionButton() {
+    const SortByHH = document.getElementById("sort_HH");
+    const SortByLL = document.getElementById("sort_LL");
+    const SortByGG = document.getElementById("sort_GG");
+
+    SortByHH.addEventListener('click', sortHhPrice);
+    SortByGG.addEventListener('click', sortGgPrice);
+    SortByLL.addEventListener('click', sortLlPrice);
+}
+
+function AddEventToSideBarButton(){
+    const MainFrame = document.querySelector(".main__ForMap");
+    const SideBarButton = MainFrame.querySelector(".main__SearchBarHideButton");
+    SideBarButton.addEventListener('click', (e) => MoveSideBar(e));
+}
+
+function AddEventToDetailTabButton(){
+    document.getElementsByClassName("main__GSTdetailCloseButton")[0].addEventListener('click', closeDetailTab);
+}
+
+function AddEventToResetButton(){
+    const SearchFromRouteButton = document.querySelector(".main__SearchFromRoute");
+    SearchFromRouteButton.addEventListener('click', (e) => MoveToinitialPoint(e));
+}
+export { initTmap,map};
